% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dremstimate.R
\name{drem_statstack}
\alias{drem_statstack}
\title{function to arrange remstats array into a stat stack for glm for the duration rem model}
\usage{
drem_statstack(
  edgelist,
  start_stats,
  end_stats,
  dur_undirected = FALSE,
  reh_undirected = FALSE,
  engaged_stat = FALSE,
  engaged_directed = TRUE
)
}
\arguments{
\item{edgelist}{data.frame with columns in order (start_time, sender, receiver, end_time)}

\item{start_stats}{remstats object for start stats}

\item{end_stats}{remstats object for end stats}

\item{dur_undirected}{(default=FALSE) boolean for if the end model should be undirected}

\item{num_actors}{number of actors}
}
\description{
function to arrange remstats array into a stat stack for glm for the duration rem model
}
\details{
start_stats has dimension (M,2D,P1) where M is the number of unique time points (over start and end of an event combined)
D = total number of dyads in the network = N*(N-1)
P1 = number of start statistics specified

end_stats has dimension (M,2D,P2) where P2 = number of end statistics specified

This function allows multiple start and end events to occur simultaneously, but only once per unique dyad in an interval
}
\keyword{internal}
