% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dremstimate.R
\name{dur.remstats}
\alias{dur.remstats}
\title{function to compute preliminary stats for the duration rem using remstats 
The stats generated from this function still require processing to be used to estimate the model}
\usage{
dur.remstats(
  start_effects,
  end_effects,
  edgelist,
  actors = NULL,
  psi = 0,
  memory = "full",
  memory_value = Inf,
  dur_undirected = FALSE
)
}
\arguments{
\item{start_effects}{formula object for remstats, used to compute the start statistics}

\item{end_effects}{formula object for remstats, used to compute the end statistics}

\item{edgelist}{data.frame with columns (start_time, sender, receiver, end_time)}

\item{actors}{character vector of actor names}

\item{memory_value}{value of half life parameter for exponential decay of memory}

\item{dur_undirected}{TRUE if riskset for the duration model needs to be undirected. See details}
}
\description{
function to compute preliminary stats for the duration rem using remstats 
The stats generated from this function still require processing to be used to estimate the model
}
\details{
dur_undirected is set to TRUE if riskset for the duration model needs to be undirected. i.e if dyad A->B is in an event, the undirected dyad (AB==BA) is at risk to end the event. This argument can be used when it is not directly observable whether the sender or receiver ended the event
}
\keyword{internal}
