% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duremstimate.R
\name{duremstimate.grid}
\alias{duremstimate.grid}
\title{Estimate hyperparameters for Duration Relational Event Model (DuREM)}
\usage{
duremstimate.grid(
  start_effects,
  end_effects,
  edgelist,
  psi_start_candidates = 1,
  psi_end_candidates = 1,
  start_undirected = FALSE,
  end_undirected = FALSE,
  memory = c("full", "decay"),
  half_life_candidates = NA,
  engaged_stat = FALSE,
  engaged_directed = FALSE,
  strip_return = TRUE,
  save_dir = NULL,
  num_cores = NULL
)
}
\arguments{
\item{start_effects}{formula object for remstats, used to compute the start statistics}

\item{end_effects}{formula object for remstats, used to compute the end statistics}

\item{edgelist}{data.frame with columns (start_time, sender, receiver, end_time)}

\item{psi_start_candidates}{numeric vector of candidate values for psi_start. default value is 1}

\item{psi_end_candidates}{numeric vector of candidate values for psi_end. default value is 1}

\item{start_undirected}{TRUE if riskset for both start and end DuREM models needs to be undirected}

\item{end_undirected}{Logical. If `TRUE`, the risk set for the end DuREM model is undirected (see Details). Default is `FALSE`.}

\item{memory}{if "full" then no memory effects are incorporated. If "decay" then decay memory effects with specified half life}

\item{half_life_candidates}{numeric vector of candidate values for half life parameters}

\item{strip_return}{logical, if TRUE then strip the heavy elements from a glm output object}

\item{save_dir}{character, local directory where to save fitted candidate model files}

\item{num_cores}{integer, number of cores to use for parallelized runs. If 1 or NULL, the grid search is sequential. default value is NULL}
}
\value{
list with element \code{loglik}, a matrix or array of loglikelihood of fitted DuREM candidate models and \code{mle}, a vector of candidate values (psi_start,psi_end(,half_life)) with maximum likelihood
}
\description{
Function to estimate the hyperparameters for duration relational event model (DuREM) using grid search
}
\details{
\code{end_undirected} is set to \code{TRUE} if riskset for the duration model needs to be undirected. i.e if dyad A->B is in an event, the undirected dyad (AB==BA) is at risk to end the event. This argument can be used when it is not directly observable whether the sender or receiver ended the event

A list of available effects for the start and end models of DuREM can be obtained with \code{\link[remstats:tie_effects]{remstats::tie_effects()}} and
for a list of undirected effects \code{\link[remstats:tie_effects]{remstats::tie_effects(directed = FALSE)}}
}
\examples{

# Define effects for the start and end model of DuREM
start_effects <- ~ 1 + remstats::inertia(scaling = "std") + remstats::reciprocity(scaling = "std")
end_effects <- ~ 1 + remstats::outdegreeSender(scaling = "std")

# Fit a DuREM model
durem::duremstimate.grid(start_effects, end_effects, dat$edgelist)

}
